+++ frontend/app/components/templates/FacePaymentPage/FacePaymentPage.tsx
@@ -136,10 +136,14 @@ export const FacePaymentPage = ({ id, name, long_description }: Props) => {
         {currentStep === 1 && (
           <div className={styles.welcome}>
             <h2>Welcome to {name} Demo App</h2>
-            <p>
-              Please access this demo via smartphone or any device with at least
-              HD camera resolution for better performance and experience
-            </p>
+            <div className={styles.specialInstruction}>
+              <p className={styles.desc}>
+                Please access this demo via <strong>smartphone</strong> or any
+                device with at least <strong>HD camera</strong> resolution for{' '}
+                <strong>better performance</strong> and{' '}
+                <strong>experience</strong>
+              </p>
+            </div>
             <Button
               color={Color.Primary}
               onClick={() => {