+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - [FRONTEND] Add solution partner cards
 - [FRONTEND] Add message to tell is the face occluded or not
 - [FRONTEND] Add Custom popup for Partner's Solution
+- [FRONTEND] Add Compared photo in face match with enrollment results
 - [BACKEND] Add 5 new seed data for partner solution service: Road Traffic Monitoring, HRIS, Citizen Apps, Health Protocol Enforcement (Covid Solution), AML / PEP
 - [BACKEND] Add new service type: `solution-partner`
 - [BACKEND] Add new controller to get services with service type: `solution-partner`
@@ -20,6 +21,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - [FRONTEND] Change analytics & FOA upload limit
+- [FRONTEND] Change instruction for analytics, face occlusion and attributes, ekyc, and face payment
 - [FRONTEND] Change face occlusion & attribute example one
 - [FRONTEND] Modified solution card stylings
 - [FRONTEND] Change new-innovations layout and style in home page